修改message_id长度问题

+++ app/src/main/java/com/qwe7002/telegram_sms/message_json.java
@@ -3,7 +3,7 @@ package com.qwe7002.telegram_sms;
 class message_json {
     @SuppressWarnings("unused")
     final boolean disable_web_page_preview = true;
-    int message_id;
+    long message_id;
     String parse_mode;
     String chat_id;
     String text;
