+++ app/src/main/java/com/qwe7002/telegram_sms/sms_send_receiver.java
@@ -43,7 +43,7 @@ public class sms_send_receiver extends BroadcastReceiver {
         final message_json request_body = new message_json();
         request_body.chat_id = chat_id;
         String request_uri = public_func.get_url(bot_token, "sendMessage");
-        int message_id = Integer.parseInt(Objects.requireNonNull(extras.getString("message_id")));
+        long message_id = Long.parseLong(Objects.requireNonNull(extras.getString("message_id")));
         if (message_id != -1) {
             Log.d(public_func.log_tag, "sms_send_receiver: Find the message_id and switch to edit mode.");
             request_uri = public_func.get_url(bot_token, "editMessageText");
